worker/requirements.txt
supabase==2.11.0
loguru==0.7.3
python-dotenv==1.0.0
requests==2.32.3
pandas==2.2.3
fpdf==1.7.2
XlsxWriter==3.2.0
reportlab==4.0.7
flask==3.0.0